Tests moved to test file, bugs fixed
